feat: 전체 코드 리팩터링
ScheduledTask / BusLocation 결합도 낮추기
OpenAPI 템플릿화
ScheduledTask 에서 UserRequest 바탕으로 요청을 하도록 변경

+++ bus-seat/src/main/java/org/kernel360/busseat/route_station/entity/BusRouteStationEntity.java
@@ -33,7 +33,7 @@ public class BusRouteStationEntity {
 	private Long stationId;
 
 	@Column(name = "station_seq", nullable = false)
-	private Short stationSeq;
+	private Integer stationSeq;
 
 	@Column(name = "station_name", nullable = false)
 	private String stationName;
